+++ R/graph.R
@@ -10,39 +10,30 @@
 #' @param prefix string indicating columns containing clustering information
 #' @param count_filter count threshold for filtering edges in the clustering
 #' graph
-#' @param prop_filter proportion threshold for filtering edges in the clustering
-#' graph
-#' @param node_colour either a value indicating a colour to use for all nodes or
-#' the name of a metadata column to colour nodes by
-#' @param node_colour_aggr if `node_colour` is a column name than a function to
-#' aggregate that column for samples in each cluster
-#' @param node_size either a numeric value giving the size of all nodes or the
-#' name of a metadata column to use for node sizes
-#' @param node_size_aggr if `node_size` is a column name than a function to
-#' aggregate that column for samples in each cluster
-#' @param node_alpha either a numeric value giving the alpha of all nodes or the
-#' name of a metadata column to use for node transparency
-#' @param node_alpha_aggr if `node_size` is a column name than a function to
-#' aggregate that column for samples in each cluster
+#' @param prop_filter in proportion threshold for filtering edges in the
+#' clustering graph
+#' @param node_aes_list nested list containing node aesthetics
 #'
-#' @return [igraph::igraph] object containing the tree graph
+#' @return [tidygraph::tbl_graph] object containing the tree graph
 #'
 #' @importFrom dplyr %>%
 #' @importFrom rlang .data
 build_tree_graph <- function(clusterings, prefix, count_filter, prop_filter,
-                             metadata, node_colour, node_colour_aggr,
-                             node_size, node_size_aggr, node_alpha,
-                             node_alpha_aggr) {
+                             metadata, node_aes_list) {
 
-    nodes <- get_tree_nodes(clusterings, prefix, metadata, node_colour,
-                            node_colour_aggr, node_size, node_size_aggr,
-                            node_alpha, node_alpha_aggr)
+    nodes <- get_tree_nodes(clusterings, prefix, metadata, node_aes_list)
 
     edges <- get_tree_edges(clusterings, prefix) %>%
         dplyr::filter(.data$count > count_filter) %>%
-        dplyr::filter(.data$proportion > prop_filter)
+        dplyr::filter(.data$in_prop > prop_filter)
 
-    graph <- igraph::graph_from_data_frame(edges, vertices = nodes)
+    graph <- tidygraph::tbl_graph(nodes = nodes, edges = edges)
+
+    # Convert resolution to factor
+    igraph::vertex_attr(graph)[[prefix]] <-
+        factor(as.numeric(igraph::vertex_attr(graph)[[prefix]]))
+
+    graph <- store_node_aes(graph, node_aes_list, metadata)
 
     return(graph)
 }
@@ -56,23 +47,10 @@ build_tree_graph <- function(clusterings, prefix, count_filter, prop_filter,
 #' @param metadata data.frame containing metadata on each sample that can be
 #' used as node aesthetics
 #' @param prefix string indicating columns containing clustering information
-#' @param node_colour either a value indicating a colour to use for all nodes or
-#' the name of a metadata column to colour nodes by
-#' @param node_colour_aggr if `node_colour` is a column name than a function to
-#' aggregate that column for samples in each cluster
-#' @param node_size either a numeric value giving the size of all nodes or the
-#' name of a metadata column to use for node sizes
-#' @param node_size_aggr if `node_size` is a column name than a function to
-#' aggregate that column for samples in each cluster
-#' @param node_alpha either a numeric value giving the alpha of all nodes or the
-#' name of a metadata column to use for node transparency
-#' @param node_alpha_aggr if `node_size` is a column name than a function to
-#' aggregate that column for samples in each cluster
+#' @param node_aes_list nested list containing node aesthetics
 #'
 #' @return data.frame containing node information
-get_tree_nodes <- function(clusterings, prefix, metadata, node_colour,
-                           node_colour_aggr, node_size, node_size_aggr,
-                           node_alpha, node_alpha_aggr) {
+get_tree_nodes <- function(clusterings, prefix, metadata, node_aes_list) {
 
     nodes <- lapply(colnames(clusterings), function(res) {
         clustering <- clusterings[, res]
@@ -88,19 +66,9 @@ get_tree_nodes <- function(clusterings, prefix, metadata, node_colour,
             node_data <- list(node_name, res_clean, cluster, size)
             names(node_data) <- c("node", prefix, "cluster", "size")
 
-            if (node_colour %in% colnames(metadata)) {
-                clust_meta <- metadata[is_cluster, node_colour]
-                node_data[node_colour] <- node_colour_aggr(clust_meta)
-            }
-
-            if (node_size %in% colnames(metadata)) {
-                clust_meta <- metadata[is_cluster, node_size]
-                node_data[node_size] <- node_size_aggr(clust_meta)
-            }
-
-            if (node_alpha %in% colnames(metadata)) {
-                clust_meta <- metadata[is_cluster, node_alpha]
-                node_data[node_alpha] <- node_alpha_aggr(clust_meta)
+            for (aes in node_aes_list) {
+                node_data <- aggr_metadata(node_data, aes[[1]], aes[[2]],
+                                           metadata, is_cluster)
             }
 
             node_data <- data.frame(node_data, stringsAsFactors = FALSE)
@@ -113,7 +81,10 @@ get_tree_nodes <- function(clusterings, prefix, metadata, node_colour,
     })
 
     nodes <- do.call("rbind", nodes)
-    nodes[, prefix] <- factor(nodes[, prefix])
+
+    stabilities <- calc_sc3_stability(clusterings)
+
+    nodes$sc3_stability <- stabilities[, 3]
 
     return(nodes)
 }
@@ -129,7 +100,7 @@ get_tree_nodes <- function(clusterings, prefix, metadata, node_colour,
 #' @return data.frame containing edge information
 #'
 #' @importFrom dplyr %>%
-#' @importFrom rlang .data
+#' @importFrom rlang .data :=
 get_tree_edges <- function(clusterings, prefix) {
 
     res_values <- colnames(clusterings)
@@ -156,15 +127,15 @@ get_tree_edges <- function(clusterings, prefix) {
 
             to_size <- sum(is_to)
 
-            trans_prop <- trans_count / to_size
+            in_prop <- trans_count / to_size
 
-            return(c(trans_count, trans_prop))
+            return(c(trans_count, in_prop))
         })
 
         from_tos$from_res <- as.numeric(gsub(prefix, "", from_res))
         from_tos$to_res <- as.numeric(gsub(prefix, "", to_res))
         from_tos$count <- transitions[1, ]
-        from_tos$proportion <- transitions[2, ]
+        from_tos$in_prop <- transitions[2, ]
 
         return(from_tos)
     })
@@ -174,8 +145,66 @@ get_tree_edges <- function(clusterings, prefix) {
                                          "C", .data$from_clust)) %>%
         dplyr::mutate(to_node = paste0(prefix, .data$to_res,
                                        "C", .data$to_clust)) %>%
-        dplyr::select(.data$from_node, .data$to_node, dplyr::everything())
+        dplyr::select(.data$from_node, .data$to_node, dplyr::everything()) %>%
+        dplyr::rename(!!as.name(paste0("from_", prefix)) := .data$from_res,
+                      !!as.name(paste0("to_", prefix)) := .data$to_res)
 
     return(edges)
 
+}
+
+
+#' Aggregate metadata
+#'
+#' Aggregate a metadata column to get a summarized value for a cluster node
+#'
+#' @param node_data data.frame containing information about a set of cluster
+#' nodes
+#' @param col_name the name of the metadata column to aggregate
+#' @param col_aggr string naming a function used to aggregate the column
+#' @param metadata data.frame providing metadata on samples
+#' @param is_cluster logical vector indicating which rows of metadata are in the
+#' node to be summarized
+#'
+#' @return data.frame with aggregated data
+aggr_metadata <- function(node_data, col_name, col_aggr, metadata,
+                          is_cluster) {
+
+    if (col_name %in% colnames(metadata)) {
+        clust_meta <- metadata[is_cluster, col_name]
+        col_aggr_fun <- match.fun(col_aggr)
+        aggr_col_name <- paste0(col_aggr, "_", col_name)
+        node_data[aggr_col_name] <- col_aggr_fun(clust_meta)
+    }
+
+    return(node_data)
+}
+
+
+#' Store node aesthetics
+#'
+#' Store the names of node attributes to use as aesthetics as graph attributes
+#'
+#' @param graph graph to store attributes in
+#' @param node_aes_list nested list containing node aesthetics
+#' @param metadata data.frame containing metadata that can be used as aesthetics
+#'
+#' @return graph with additional attributes
+store_node_aes <- function(graph, node_aes_list, metadata) {
+
+    for (node_aes_name in names(node_aes_list)) {
+
+        node_aes <- node_aes_list[[node_aes_name]]$value
+        node_aes_value <- node_aes
+        node_aes_aggr <- node_aes_list[[node_aes_name]]$aggr
+
+        if (node_aes %in% colnames(metadata)) {
+            node_aes_value <- paste0(node_aes_aggr, "_", node_aes)
+        }
+
+        graph <- igraph::set_graph_attr(graph, paste0("node_", node_aes_name),
+                                        node_aes_value)
+    }
+
+    return(graph)
 }